requirements.txt
django-crispy-forms==1.7.2
pytz==2018.5
